requirements.txt
python-dateutil>=2.8.1
requests==2.27.1
pandas==1.4.0
plotly==5.5.0
streamlit==1.4.0
lxml==4.7.1
scrapy >= 1.7.3
SQLAlchemy>=1.3.6
boto3>=1.20.42
botocore>=1.23.42
numpy~=1.22.1
python-dotenv~=0.19.2
itemadapter~=0.4.0